refactor(blackjack): GameResult를 도메인 성격으로 분류
