+++ src/main/java/blackjack/ApplicationRunner.java
@@ -1,8 +1,6 @@
 package blackjack;
 
 import blackjack.controller.BlackjackController;
-import blackjack.domain.card.CardFactory;
-import blackjack.domain.card.Cards;
 
 public class ApplicationRunner {
