+++ src/test/java/blackjack/GameResultTest.java
@@ -1,5 +1,6 @@
 package blackjack;
 
+import blackjack.domain.GameResult;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardNumber;
 import blackjack.domain.card.Symbol;
